+++ TagsCloudVisualization/Exceptions/SpiralPointGeneratorException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TagsCloudVisualization
+{
+    public class SpiralPointGeneratorException : Exception
+    {
+        public SpiralPointGeneratorException() : base("Failed searching suitable point") {}
+    }
+}
